pyproject.toml
[project]
name = "diary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "lockfile>=0.12.2",
    "python-daemon>=3.1.2",
    "tabulate>=0.9.0",
]
